add typedef uintXX_t numbers; update lexer to new token data types

+++ src/include/lexer.h
@@ -1,7 +1,9 @@
 #ifndef LEXER_H
 #define LEXER_H
 
+#include <stddef.h>
 #include <stdint.h>
+#include "numbers.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -18,26 +20,26 @@ typedef struct Lexer {
 	// If the lexer is raw, `source` will not be freed
 	bool raw;
     // Length of source code string
-    long length;
+    size_t length;
     // Array of tokens; Filled with ScanTokens(...) function
     Token **tokens;
 
     // Start index of the character for token to be created
     // When creating a new token, this will be the first character index
-    long start;
+    size_t start;
     // Current index of the character being read
     // Will be last character of newly created token
-    long current;
+    size_t current;
     // Current line number
-    long line;
+    size_t line;
     // current column number of the start of the token
-    long column;
+	size_t column;
 
     // Codepoint iterator. The characters are read as UTF-32 characters
     UIter *iter;
 
     // Timestamp seed for hashing
-    uint64_t timestamp;
+    u64 timestamp;
 
     // Reference to Pankti Core
     void *core;
