+++ src/include/token.h
@@ -1,12 +1,16 @@
 #ifndef TOKEN_H
 #define TOKEN_H
 
+#include "numbers.h"
+#include <stddef.h>
+#include <stdbool.h>
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#include <stdbool.h>
-#include <stdint.h>
+
 
 // Token Types
 typedef enum TokenType {
@@ -132,15 +136,15 @@ typedef struct Token {
     // Can be NULL
     char *lexeme;
     // Line number
-    long line;
+    size_t line;
     // Column number. Distance from first character of the line
-    long col;
+    size_t col;
     // length of the token lexeme, if lexeme is optional depends of how many
     // characters the token contains need. For example, T_EQ is 1, T_EQEQ is 2
-    long len;
-    // Hash for the token lexeme, expect for Keywords, identifiers, strings,
+    size_t len;
+    // Hash for the token lexeme. expect for Keywords, identifiers, strings,
     // numbers it will be 0 (zero)
-    uint64_t hash;
+    u64 hash;
 } Token;
 
 // Create an empty token.
